pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ftl-gen"
version = "0.1.0"
description = "LLM-powered FTL mod generator with custom sprite generation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "FTL Gen" }
]
dependencies = [
    "anthropic>=0.25.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
ftl-gen = "ftl_gen.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ftl_gen"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
